Squashed commit of the following:

commit 7c2b4acee169fa2bd19e7a145c3d6d3b13e3f721
Merge: 5082876 bf74fd0
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 18:19:54 2019 -0800

    Merge branch 'US1083' into US1083_bitbucket_resolution

commit bf74fd0881484b1b99e53ccb72123f8d5c6e4c61
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 16:10:23 2019 -0800

    Removed comment pair

commit 75e949ca9a5f823efc85901b084402715fa5c3c6
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:53:28 2019 -0800

    Updated .gitignore

commit 44b64e134f96794c805d4d376be1aef1dd6d3a20
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:51:38 2019 -0800

    Merged with master on 12-19-19

commit a71fba8bf38a6a2dfa2f6f29018c82f656bd71d9
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:50:59 2019 -0800

    Merged with master on 12-19-19

commit 71d91de685491e61dbd8676a6b3d9b3e50083958
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:49:41 2019 -0800

    Merged with master on 12-19-19

commit b88b059aac20ba46bfcacb83978168b538741817
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:49:10 2019 -0800

    Merged with master on 12-19-19. Added destructor, updated lookupCD method to not accept parameter because of new try catch block. Added method member and associated methods for control surface boolean (this used to be passed into lookupCD)

commit e4b8eef9e42f8f0c900a631edcc9a1d704a25f74
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:47:23 2019 -0800

    Added exception classes

commit a9b102abe8aa0ff67d6f397999b864499d4f8074
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:47:04 2019 -0800

    Updated gitignore

commit 637403c1d8479cdcbf354125a30306c043b18605
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Dec 19 13:46:13 2019 -0800

    Merged with master on 12-19-19

commit 57fb6fd67705d12ef16dafd289b9fcf3b0e15529
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Thu Oct 10 12:49:37 2019 -0700

    Removed conditionAngle method (obsolete)

commit f1ea3fb0b094a63a34084cea86eb2997de10964a
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 18:21:47 2019 -0700

    Changed all floats to double, added tests for lookupCD receiving true

commit d3724f8e62e4e0460387f13d516cdf83b359bd98
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 18:20:34 2019 -0700

    Changed all float types to double.

commit 6de824feefb4cad7854b30a68cefc3a52275d68f
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 12:53:26 2019 -0700

    Fixed some formatting issues.

commit 19e2eb0eca4a0236ff8f7e89b800d455f08a7e92
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 12:38:44 2019 -0700

    Updated test to include code to calculate rotated lift and drag for non-control surface cases.

commit 06154e65dbb89a743445dd0d15c108248615878e
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 12:36:31 2019 -0700

    Removed calculateLiftDragValues method. Extended lookupCD to take boolean isForControlSurface

commit 757ae8aa7f97bcf376b0e9b6d2da44fe1d047fc9
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 12:26:30 2019 -0700

    Removed calculateLiftDragModelValues method, added parameter isForControlSurface to lookupCD (drag coefficient now calculated differently according to whether or not it is for a control surface)

commit df9e251ddffa1159cc7ec4b6f9ab98ec1a26cde5
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 12:24:29 2019 -0700

    Updated data to include truths with rotated lift and drag.

commit 3102ae2f423e67241d6e64f03fa973607c732224
Author: Ihimu Ukpo <iukpo@swiftengineering.com>
Date:   Wed Oct 2 12:23:24 2019 -0700

    Updated gitignore

+++ include/avionics_sim/Bilinear_interp.hpp
@@ -45,7 +45,7 @@ namespace avionics_sim
           /// \param[in]  yv  Reference to Y values of the 2D LUT
           /// \param[in]  zv  Reference to Z values of the 2D LUT
           /// X values and Y values must be in ascending order. 
-          Bilinear_interp(const std::vector<std::vector<float>>& xv, const std::vector<float>& yv, const std::vector<std::vector<float>>& zv);
+          Bilinear_interp(const std::vector<std::vector<double>>& xv, const std::vector<double>& yv, const std::vector<std::vector<double>>& zv);
 
           // Function to perform 1D (Linear) interpolation.     
           ///
@@ -61,7 +61,7 @@ namespace avionics_sim
           ///
           /// \return      Returns InterpResult enum containing potential interpolation errors. 
           /// X values must be ascending
-          InterpResult interpolate(const std::vector<float> &xv, const std::vector<float> &yv, float x, float* const y);
+          InterpResult interpolate(const std::vector<double> &xv, const std::vector<double> &yv, double x, double* const y);
           
 
 
@@ -81,7 +81,7 @@ namespace avionics_sim
           ///
           /// \return      Returns InterpResult enum containing potential interpolation errors. 
           /// X and Y values must be ascending
-          InterpResult interpolate2D(const std::vector<std::vector<float>>& xv, const std::vector<float>& yv, const std::vector<std::vector<float>>& zv, float x, float y, float* const z);
+          InterpResult interpolate2D(const std::vector<std::vector<double>>& xv, const std::vector<double>& yv, const std::vector<std::vector<double>>& zv, double x, double y, double* const z);
 
           ///
           /// \brief      Performs 2D interpolations on LUT based on x, y, and z values stored within the class. 
@@ -96,7 +96,7 @@ namespace avionics_sim
           ///
           /// \return     Returns InterpResult enum containing potential interpolation errors. 
           /// X and Y values must be ascending
-          InterpResult interpolate2D(float x, float y, float* const z);
+          InterpResult interpolate2D(double x, double y, double* const z);
 
 
           
@@ -107,11 +107,11 @@ namespace avionics_sim
           ///            withing a LUT. 
           ///
           /// \param[in]  inputVals The input string to be processed. 
-          /// \param[out] outputVect A pointer to std::vector<float> containing the LUT elements.
+          /// \param[out] outputVect A pointer to std::vector<double> containing the LUT elements.
           ///
           /// \return     A bool indicating whether the parsing was a success. 
           ///
-          static bool get1DLUTelementsFromString(const std::string& inputVals, std::vector<float>* const outputVect);
+          static bool get1DLUTelementsFromString(const std::string& inputVals, std::vector<double>* const outputVect);
 
           ///
           /// \brief      Gets a two dimensional vector of LUT elements from a string. 
@@ -119,11 +119,11 @@ namespace avionics_sim
           ///            values representing points on a 2D LUT. 
           ///
           /// \param[in]  inputVals  The input string to be processed. 
-          /// \param[out] outputVect A pointer to std::vector<std::vector<<float>> containing the 2D LUT elements.
+          /// \param[out] outputVect A pointer to std::vector<std::vector<<double>> containing the 2D LUT elements.
           ///
           /// \return     A bool indicating whether the parsing was a success. 
           ///
-          static bool get2DLUTelementsFromString(const std::string& inputVals, std::vector<std::vector<float>>* const outputVect);
+          static bool get2DLUTelementsFromString(const std::string& inputVals, std::vector<std::vector<double>>* const outputVect);
           
 
 
@@ -143,7 +143,7 @@ namespace avionics_sim
           ///
           /// \param[in]  xv     Reference to vector containing the x values. 
           ///
-          void setXVals(const std::vector<std::vector<float>>& xv);
+          void setXVals(const std::vector<std::vector<double>>& xv);
 
           ///
           /// \brief      Sets the y vals vector.
@@ -161,7 +161,7 @@ namespace avionics_sim
           ///
           /// \param[in]  yv     Reference to vector containing the y values. 
           ///
-          void setYVals(const std::vector<float>& yv);
+          void setYVals(const std::vector<double>& yv);
 
 
           ///
@@ -180,7 +180,7 @@ namespace avionics_sim
           ///
           /// \param[in]  zv     Reference to vector containing the z values. 
           /// 
-          void setZVals(const std::vector<std::vector<float>>& zv);
+          void setZVals(const std::vector<std::vector<double>>& zv);
 
           ///
           /// \brief      Gets the x vals vector.
@@ -189,7 +189,7 @@ namespace avionics_sim
           /// 
           /// \return A bool indicating whether the operation was a success.  
           ///
-          bool getX(std::vector<std::vector<float>>* const xVect); 
+          bool getX(std::vector<std::vector<double>>* const xVect); 
 
           ///
           /// \brief      Gets the y vals vector.
@@ -198,7 +198,7 @@ namespace avionics_sim
           /// 
           /// \return A bool indicating whether the operation was a success.  
           ///
-          bool getY(std::vector<float>* const yVect); 
+          bool getY(std::vector<double>* const yVect); 
 
           ///
           /// \brief      Gets the z vals vector.
@@ -207,13 +207,13 @@ namespace avionics_sim
           /// 
           /// \return A bool indicating whether the operation was a success.  
           ///
-          bool getZ(std::vector<std::vector<float>>* const zVect); 
+          bool getZ(std::vector<std::vector<double>>* const zVect); 
 
         private: 
           /// \brief Vectors to hold LUT data.
           /// Holds 2D LUT values for X and Z points
-          std::vector< std::vector<float> > xVals, zVals;
-          std::vector<float> yVals; ///< Holds Y values for LUT data.
+          std::vector< std::vector<double> > xVals, zVals;
+          std::vector<double> yVals; ///< Holds Y values for LUT data.
 
           /// Bools to keep track of which LUT data values have been added. 
           bool haveXVals, haveYVals, haveZVals;
