+++ tests/unit_tests/LiftDragCalculateDragParameterized.h
@@ -21,9 +21,9 @@ typedef struct LiftDragCalculateDragParams {
 	double rho;
 	double vInf;
 	double angle;
-	std::vector<float> LUT_alpha;
-	std::vector<float> LUT_CL;
-	std::vector<float> LUT_CD;
+	std::vector<double> LUT_alpha;
+	std::vector<double> LUT_CL;
+	std::vector<double> LUT_CD;
 }LiftDragCalculateDragParams;
 
 //Type definition for LiftDragParameterCollections.
