+++ tests/unit_tests/Test_Lift_Drag_Lookup_CD.cpp
@@ -13,9 +13,9 @@ TEST_P(LiftDragLookupCDParameterized, LiftDragUnitTestLookupCD_UnitTest) {
 	avionics_sim::Lift_drag_model ldm;
 	double inputAngle;
 	LiftDragLookupCDParams param=GetParam();
-	std::vector<float> LUT_alpha; 
-	std::vector<float> LUT_CL; 
-	std::vector<float> LUT_CD; 
+	std::vector<double> LUT_alpha; 
+	std::vector<double> LUT_CL; 
+	std::vector<double> LUT_CD; 
 	inputAngle=param.value;
 	LUT_alpha=param.LUT_alpha;
 	LUT_CL=param.LUT_CL;
