+++ tests/unit_tests/Test_Lift_Drag_Lookup_CL.cpp
@@ -12,9 +12,9 @@
 TEST_P(LiftDragLookupCLParameterized, LiftDragUnitTestLookupCL_UnitTest) {
 	avionics_sim::Lift_drag_model ldm;
 	double inputAngle;
-	std::vector<float> LUT_alpha; 
-	std::vector<float> LUT_CL; 
-	std::vector<float> LUT_CD; 
+	std::vector<double> LUT_alpha; 
+	std::vector<double> LUT_CL; 
+	std::vector<double> LUT_CD; 
 	LiftDragLookupCLParams param=GetParam();
 	inputAngle=param.value;
 	LUT_alpha=param.LUT_alpha;
